remove mlflow

+++ src/dataio.py
@@ -93,7 +93,7 @@ class Artist(torch.utils.data.Dataset):
         return len(self.data)
 
     def __getitem__(self, idx):
-        data = torch.from_numpy(np.load(self.data[idx]))
+        data = torch.from_numpy(np.load(self.data[idx]),dtype=torch.float32)
         label = self.labels[idx]
         return data,label
